src/requirements.txt
# # # # # # # # # # # # # # # # # # # #
# NewsBot, a journalism tool
# Copyright 2020 Carter Pape
# 
# See file LICENSE for licensing terms.
# # # # # # # # # # # # # # # # # # # #

scrapy
pape
twisted
python-magic
python-dotenv
mysql-connector-python
requests
dateparser
metadata_parser
pylint
